+++ integrations/github/github/webhook/events.py
@@ -13,4 +13,4 @@ REPOSITORY_DELETE_EVENTS = ["archived", "deleted"]
 ALL_EVENTS = REPOSITORY_UPSERT_EVENTS + REPOSITORY_DELETE_EVENTS
 
 
-WEBHOOK_CREATE_EVENTS = ["repository"]
+WEBHOOK_CREATE_EVENTS = ["repository", "create", "delete", "push"]